Added a feature for Session Notes. Session Notes is a free form text field associated with each session. It can be used to take notes about the conditions, car setup, location, etc.
Added a fifth tab for Session Notes.

On the Session Notes tab the filters for Driver and Car appear, along with the Date Range Filters.

Selecting a Session will display the notes for that session. There is an edit button to allow the user to edit / update the session notes.