Added a Location feature.
The user can add, edit, and delete Locations on the driver-car-manager.tsx control.

Each Session is associated with a Location

Deleting a Location will delete all Sessions associated with that Location.

In the Session Configuration section of lap-timer.tsx the user is required to select a Location after Selecting a Driver and a Car.

In the Session Configuration section of lap-timer.tsx the user has the ability to create a "New Location" just like they can create a "New Driver" or "New Car"

Location data should be stored into and retrieved from the database.